requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-admin-thumbnails==0.2.6
django-ckeditor==6.3.1
django-filter==21.1
django-jalali==6.0.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-taggit==3.1.0
ghasedak==0.1.7
ghasedakpack==0.1.13
idna==3.3
jdatetime==4.1.0
Pillow==10.1.0
pyotp==2.9.0
requests==2.27.1
setuptools==78.1.0
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.2
suds==1.1.1
tzdata==2021.5
urllib3==1.26.8
wheel==0.45.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
